+++ src/lib/branding.ts
@@ -1,5 +1,5 @@
 export const APP_NAME = 'Torfin'
-export const APP_VERSION = '1.0.0-beta'
+export const APP_VERSION = '1.0.0-beta.4'
 export const APP_USER_AGENT = `${APP_NAME}/${APP_VERSION}`
 export const APP_STORAGE_PREFIX = 'torfin'
 export const JELLYFIN_CLIENT = `${APP_NAME}`